user can get and delete their own games

+++ src/main/kotlin/hh/project/discgolf/services/GameService.kt
@@ -1,13 +1,17 @@
 package hh.project.discgolf.services
 
 import hh.project.discgolf.entities.Game
+import hh.project.discgolf.entities.User
 import hh.project.discgolf.repositories.GameRepository
+import hh.project.discgolf.repositories.UserRepository
+import jakarta.persistence.Id
 import org.springframework.data.crossstore.ChangeSetPersister.NotFoundException
+import org.springframework.security.core.Authentication
 import org.springframework.stereotype.Service
 import java.time.LocalDateTime
 
 @Service
-class GameService(private val gameRepository: GameRepository) {
+class GameService(private val gameRepository: GameRepository, private val userRepository: UserRepository) {
 
 
     fun getAllGames(): List<Game> = gameRepository.findAll()
@@ -18,12 +22,15 @@ class GameService(private val gameRepository: GameRepository) {
 
     fun createGame(game: Game): Game = gameRepository.save(game)
 
-    fun deleteGame(gameId: Long) {
+    fun deleteGame(gameId: Long, userId: Long, authentication: Authentication) {
+        val user = authentication.principal as User
+        val game = gameRepository.findById(gameId)
 
-
-        return if (gameRepository.existsById(gameId)) {
+        if (game.isPresent && user.userId == game.get().user?.userId) {
             gameRepository.deleteById(gameId)
-        } else throw NoSuchElementException("User doesn't exist with given id!")
+        } else {
+            throw NoSuchElementException("Game not found or you are not authorized to delete this game")
+        }
     }
 
     fun updateGame(gameId: Long, game: Game): Game {
@@ -48,4 +55,10 @@ class GameService(private val gameRepository: GameRepository) {
             gameRepository.save(updatedGame)
         } else throw NotFoundException()
     }
+
+    fun getUserGames(userId: Long): List<Game> {
+        return gameRepository.getUserGames(userId)
+    }
 }
+
+
